// frontend/src/types/api.d.ts
export interface components {
    schemas: {
// ...
        AdminProductRequest: {
            category: number;
            slug?: string;
            /** @description Price in UZS, whole units only. */
            price: number;
            is_available?: boolean;
            order?: number;
            translations: components["schemas"]["ProductTranslationRequest"][];
        };
// ...
        PaginatedAdminProductList: {
            /** @example 123 */
            count: number;
            /**
             * Format: uri
             * @example http://api.example.org/accounts/?page=4
             */
            next?: string | null;
            /**
             * Format: uri
             * @example http://api.example.org/accounts/?page=2
             */
            previous?: string | null;
            results: components["schemas"]["AdminProduct"][];
        };
// ...
        PatchedAdminProductRequest: {
            category?: number;
            slug?: string;
            /** @description Price in UZS, whole units only. */
            price?: number;
            is_available?: boolean;
            order?: number;
            translations?: components["schemas"]["ProductTranslationRequest"][];
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}